requirements.txt
json
numpy>=1.23.4
requests
tqdm
pandas>=1.5.3
gradio==4.15.0
tiktoken
rich
portalocker
torch>=2.0.1
timeout-decorator
transformers>=4.36.2
opencv-python>=4.4.0.46
typing_extensions==4.7.1
pillow
omegaconf
matplotlib
einops
sentencepiece
sty
huggingface_hub
visual_genome
pycocoevalcap
openpyxl
seaborn
tabulate
xlsxwriter